Add files via upload
This is a follow up on Randy Olson work on solving TSPs in Python: Computing the optimal road trip across the U.S.. The code below is commented at length in my blog entry on Computing The Really Optimal Tour Across The USA On The Cloud With Python

Some of the code is a derivative work of Randy Olson's code, namely the list of points to visits, and the html template used for rendering the tour on Google Map.

Can we do better than Randy, i.e. can we compute the shortest tour visiting each one of the 50 landmarks he has selected?

The answer is yes. I will use these tools to compute it.
Google Maps

No need to present this, is it? We will use it for gathering data and rendering the result on a map.
Concorde

Concorde is the best known algorithm to solve TSPs, see my post for more background on it.
CPLEX

We use CPLEX indirectly as Concorde is built on top of it. In all fairness, Concorde can also be used with an alternative solver called QSopt. Using CPLEX is faster, but th